refactor: Simplify default implementations and callback usage in UI components
- Replaced custom default implementations for various enum variants with the `#[derive(Default)]` attribute for cleaner code.
- Updated callback invocations in multiple components to remove unnecessary cloning, improving performance and readability.
- Ensured consistency across components by standardizing the handling of optional callbacks.

+++ app/src/concepts/contribution/actions.rs
@@ -65,11 +65,11 @@ pub fn initialize_mock_repository_contributions(username: &str) -> Vec<Repositor
         .iter()
         .map(|repo| RepositoryContribution {
             repository: repo.clone(),
-            commits: (repo.stars / 10) as u32,
-            prs: (repo.stars / 20) as u32,
-            reviews: (repo.stars / 30) as u32,
-            lines_added: (repo.stars * 5) as u32,
-            lines_deleted: (repo.stars * 2) as u32,
+            commits: (repo.stars / 10),
+            prs: (repo.stars / 20),
+            reviews: (repo.stars / 30),
+            lines_added: (repo.stars * 5),
+            lines_deleted: (repo.stars * 2),
             percentage: repo
                 .contributors
                 .iter()
